+++ lib/auth-backend/venue_identity.rb
@@ -2,7 +2,7 @@ require 'active_record'
 
 module Auth::Backend
   class VenueIdentity < ::ActiveRecord::Base
-    attr_accessible :venue, :venue_id, :user_id
+    attr_accessible :venue, :venue_id, :user_id, :name, :email
 
     validates :venue, inclusion: {in: %w{facebook galaxy-spiral}}