+++ spec/requests/api_spec.rb
@@ -17,6 +17,110 @@ describe "Authentication API" do
     Apps.setup_oauth_api_client_app!
   end
 
+  describe "User API" do
+    before do
+      username = @user['name']
+      password = @password
+
+      authed_client = Rack::Client.new {
+       run Rack::Client::Auth::Basic.new(APP, username, password, true)
+      }
+      response = authed_client.post("http://auth-backend.dev/api/v1/token")
+      @token = JSON.parse(response.body)['token']
+
+      VenueIdentity.destroy_all
+      @facebook_data = {
+        "venue-id" => "1234567",
+        "name" => "Peter Smith",
+        "email" => "peter@example.com"
+      }
+      @galaxy_spiral_data = {
+        "venue-id" => "asff3564",
+        "name" => "Peter S",
+        "email" => "psmith@example.com"
+      }
+
+      app = TEST_HELPERS.create_app!
+
+      authed_client = Rack::Client.new {
+        run Rack::Client::Auth::Basic.new(APP, app[:id], app[:secret], true)
+      }
+      response = authed_client.post('http://auth-backend.dev/api/v1/token/app')
+      @app_token = JSON.parse(response.body)['token']
+    end
+
+    it "can get the venue identities of a user" do
+      response = client.post("http://auth-backend.dev/api/v1/token/venue/facebook", {'Authorization' => "Bearer #{@app_token}"}, JSON.dump(@facebook_data))
+      response.status.must_equal 201
+      response = client.post("http://auth-backend.dev/api/v1/token/venue/galaxy-spiral", {'Authorization' => "Bearer #{@app_token}"}, JSON.dump(@galaxy_spiral_data))
+      response.status.must_equal 201
+      VenueIdentity.count.must_equal 2
+      VenueIdentity.all.each {|vi| vi.user = User.find(@user['id']); vi.save!}
+
+      uuid = @user['uuid']
+
+      response = client.get "http://auth-backend.dev/api/v1/users/#{uuid}/identities", {'Authorization' => "Bearer #{@token}"}
+      response.status.must_equal 200
+
+      data = JSON.parse(response.body)
+      data['uuid'].must_equal uuid
+      data['venues'].must_equal(
+        'facebook' => {'name' => @facebook_data['name'],
+                       'id' => @facebook_data['venue-id']},
+        'galaxy-spiral' => {'name' => @galaxy_spiral_data['name'],
+                            'id' => @galaxy_spiral_data['venue-id']}
+      )
+    end
+
+    it "can get the venue identities of a bunch of users" do
+      response = client.post("http://auth-backend.dev/api/v1/token/venue/facebook", {'Authorization' => "Bearer #{@app_token}"}, JSON.dump(@facebook_data))
+      response.status.must_equal 201
+      response = client.post("http://auth-backend.dev/api/v1/token/venue/galaxy-spiral", {'Authorization' => "Bearer #{@app_token}"}, JSON.dump(@galaxy_spiral_data))
+      response.status.must_equal 201
+      VenueIdentity.count.must_equal 2
+      VenueIdentity.all.each {|vi| vi.user = User.find(@user['id']); vi.save!}
+
+      @user2 = TEST_HELPERS.create_user!(name: 'Sam', email: 'sam@example.com', password: @password, admin: 'false')
+      response = client.post("http://auth-backend.dev/api/v1/token/venue/galaxy-spiral", {'Authorization' => "Bearer #{@app_token}"}, JSON.dump({
+        'venue-id' => '438957380',
+        'name' => 'Sam The Man',
+        'email' => 'samman@example.com'
+      }))
+      response.status.must_equal 201
+      vi = VenueIdentity.where('venue_id' => '438957380').first
+      vi.user_id = @user2['id']
+      vi.save!
+
+      @user3 = TEST_HELPERS.create_user!(name: 'Jacko', email: 'jacko@example.com', password: @password, admin: 'false')
+
+      response = client.get "http://auth-backend.dev/api/v1/users/batch/identities", {'Authorization' => "Bearer #{@token}"}, JSON.dump([@user['uuid'], @user2['uuid'], @user3['uuid']])
+      response.status.must_equal 200
+
+      data = JSON.parse(response.body)
+      data.must_equal({
+        @user['uuid'] => {
+          'uuid' => @user['uuid'],
+          'venues' => {
+            'facebook' => {'name' => @facebook_data['name'],
+                           'id' => @facebook_data['venue-id']},
+             'galaxy-spiral' => {'name' => @galaxy_spiral_data['name'],
+                                 'id' => @galaxy_spiral_data['venue-id']}
+          }
+        },
+        @user2['uuid'] => {
+          'uuid' => @user2['uuid'],
+          'venues' => {
+            'galaxy-spiral' => {'name' => 'Sam The Man', 'id' => '438957380'}
+          }
+        },
+        @user3['uuid'] => {
+          'uuid' => @user3['uuid'],
+          'venues' => {}
+        }
+      })
+    end
+  end
+
   describe "OAuth API" do
     it "can't get a token without credentials" do
       response = client.post("http://auth-backend.dev/api/v1/token")